serving/fn2/requirements.txt
parliament-functions==0.1.0
scikit-learn==0.24.2
pandas
joblib
requests
pytest